# [core] Update writer/reader benchmark for bloom filter